fixed authentication

+++ controllers/contacts.js
@@ -2,34 +2,73 @@ const { Contact } = require('../models/contact');
 
 const { HttpError, ctrlWrapper } = require('../helpers');
 
-const getAll = async (req, res) => {
-  const allContacts = await Contact.find();
-  res.status(200).json(allContacts);
+const getAll = async (req, res, next) => {
+  try {
+    const { _id: owner } = req.user;
+    const { page = 1, limit = 20, ...filters } = req.query;
+    const skip = (page - 1) * limit;
+    const query = { owner, ...filters };
+
+    let contacts;
+
+    if (Object.keys(filters).length > 0) {
+      contacts = await Contact.find(query, '-__v', { skip, limit }).populate(
+        'owner',
+        'email subscription'
+      );
+    } else {
+      contacts = await Contact.find({ owner }, '-__v', { skip, limit }).populate(
+        'owner',
+        'email subscription'
+      );
+    }
+
+    res.status(200).json(contacts);
+  } catch (error) {
+    next(error);
+  }
 };
 
 const getContact = async (req, res) => {
   const { contactId } = req.params;
+  const { _id: currentUser } = req.user;
   const oneContact = await Contact.findById(contactId);
 
   if (!oneContact) {
     throw HttpError(404, 'Not found');
   }
-  res.status(200).json(oneContact);
+  const { owner } = oneContact;
+  if (!owner || owner.toString() !== currentUser.toString()) {
+    res.status(403).json({ message: 'Access is denied' });
+  } else {
+    res.status(200).json(oneContact);
+  }
 };
 
 const createContact = async (req, res) => {
-  const newContact = await Contact.create(req.body);
+  const { _id: owner } = req.user;
+  const newContact = await Contact.create({ ...req.body, owner });
+
   res.status(201).json(newContact);
 };
 
 const deleteContact = async (req, res) => {
   const { contactId } = req.params;
-  const deletedContact = await Contact.findByIdAndDelete(contactId);
-  if (!deletedContact) {
+  const { _id: currentUser } = req.user;
+  let oneContact = await Contact.findById(contactId);
+
+  if (!oneContact) {
     throw HttpError(404, 'Not found');
   }
 
-  res.status(200).json('Contact deleted');
+  const { owner } = oneContact;
+
+  if (!owner || owner.toString() !== currentUser.toString()) {
+    res.status(403).json({ message: 'Access is denied' });
+  } else {
+    oneContact = await Contact.findByIdAndDelete(contactId);
+    res.status(200).json({ message: 'Contact deleted' });
+  }
 };
 
 const renewContact = async (req, res) => {
@@ -38,24 +77,45 @@ const renewContact = async (req, res) => {
   }
 
   const { contactId } = req.params;
-  const updatedContact = await Contact.findByIdAndUpdate(contactId, req.body, { new: true });
-  if (!updatedContact) {
+  const { _id: currentUser } = req.user;
+  let oneContact = await Contact.findById(contactId);
+
+  if (!oneContact) {
     throw HttpError(404, 'Not found');
   }
-  res.status(200).json(updatedContact);
+
+  const { owner } = oneContact;
+
+  if (!owner || owner.toString() !== currentUser.toString()) {
+    res.status(403).json({ message: 'Access is denied' });
+  } else {
+    oneContact = await Contact.findByIdAndUpdate(contactId, req.body, { new: true });
+    res.status(200).json(oneContact);
+  }
 };
 
 const updateStatusContact = async (req, res) => {
   if (Object.keys(req.body).length === 0) {
-    throw HttpError(400, 'missing field favorite');
+    throw HttpError(400, 'Missing favorite field');
   }
 
   const { contactId } = req.params;
-  const updatedContact = await Contact.findByIdAndUpdate(contactId, req.body, { new: true });
-  if (!updatedContact) {
+  const { _id: currentUser } = req.user;
+
+  let oneContact = await Contact.findById(contactId);
+
+  if (!oneContact) {
     throw HttpError(404, 'Not found');
   }
-  res.status(200).json(updatedContact);
+
+  const { owner } = oneContact;
+
+  if (!owner || owner.toString() !== currentUser.toString()) {
+    res.status(403).json({ message: 'Access is denied' });
+  } else {
+    oneContact = await Contact.findByIdAndUpdate(contactId, req.body, { new: true });
+    res.status(200).json(oneContact);
+  }
 };
 
 module.exports = {
